+++ src/stringutils.c
@@ -178,7 +178,7 @@ char *strjoin(char **strings, unsigned int count, char *join) {
 }
 
 
-void strformat(char *str, int length){
+void strformat(char *str, size_t length){
     if(strlen(str) <= 0 || length <= 0) return;
 
     if(strlen(str) <= length){
